Declare context for emmc for common
Declare context for emmc and give permission for RIDL2.0. RIDL2.0 needs to
be able to write to sysfs dload and dload_mode to enable emmc crashdumps
and mini crashdumps.

It also needs to be able to read/write to the rawdump device block to copy
the crashdump for post processing and mark it as read.

Initial support was for 8937, 8996, 8953, 8976, and 8998 for full dumps
and then support was added for sdm660/670 for minidumps.
This solution moves the target specific to common, which was needlessly
getting replicated for new targets.

Change-Id: I6a0a692a52c4a4a51716f844ea5361af2a2d12da
CRs-Fixed: 2195014